feat: implement collect franchises daily bonus

+++ CityVilleDotnet.Domain/Entities/Franchise.cs
@@ -5,7 +5,7 @@ public class Franchise
     public int Id { get; set; }
     public string FranchiseType { get; set; }
     public string FranchiseName { get; set; }
-    public int TimeLastCollected { get; set; }
+    public long TimeLastCollected { get; set; }
     public List<FranchiseLocation> Locations { get; set; } = [];
 
     public Franchise(string franchiseType, string franchiseName)
